refactor: 프로젝트 상세 내부 라우팅 및 컴포넌트 구조 정리(#203)
